+++ rexp2.c
@@ -10,7 +10,7 @@ the GNU General Public License, version 2, 1991.
 ********************************************/
 
 /*
- * $MawkId: rexp2.c,v 1.17 2010/05/07 00:48:50 tom Exp $
+ * $MawkId: rexp2.c,v 1.18 2010/05/07 22:08:36 tom Exp $
  * @Log: rexp2.c,v @
  * Revision 1.3  1993/07/24  17:55:12  mike
  * more cleanup
@@ -219,7 +219,7 @@ REtest(char *str,		/* string to test */
 
     /* handle the easy case quickly */
     if ((m + 1)->s_type == M_ACCEPT && m->s_type == M_STR) {
-	return str_str(s, len, m->s_data.str, m->s_len) != (char *) 0;
+	return str_str(s, len, m->s_data.str, (size_t) m->s_len) != (char *) 0;
     } else {
 	u_flag = U_ON;
 	stackp = RE_run_stack_empty;
@@ -240,7 +240,7 @@ REtest(char *str,		/* string to test */
 
     switch (m->s_type + u_flag) {
     case M_STR + U_OFF + END_OFF:
-	if (strncmp(s, m->s_data.str, m->s_len))
+	if (strncmp(s, m->s_data.str, (size_t) m->s_len))
 	    goto refill;
 	s += m->s_len;
 	m++;
@@ -254,7 +254,7 @@ REtest(char *str,		/* string to test */
 	goto reswitch;
 
     case M_STR + U_ON + END_OFF:
-	if (!(s = str_str(s, (size_t) (str_end - s), m->s_data.str, m->s_len)))
+	if (!(s = str_str(s, (size_t) (str_end - s), m->s_data.str, (size_t) m->s_len)))
 	    goto refill;
 	push(m, s + 1, sp, U_ON);
 	s += m->s_len;
@@ -264,7 +264,7 @@ REtest(char *str,		/* string to test */
 
     case M_STR + U_ON + END_ON:
 	t = (str_end - s) - m->s_len;
-	if (t < 0 || memcmp(s + t, m->s_data.str, m->s_len))
+	if (t < 0 || memcmp(s + t, m->s_data.str, (size_t) m->s_len))
 	    goto refill;
 	s = str_end;
 	m++;
